+++ cli/tauri.js/api-src/fs.ts
@@ -18,7 +18,7 @@ export enum BaseDirectory {
   Template,
   Video,
   Resource,
-  App,
+  App
 }
 
 export interface FsOptions {
@@ -45,15 +45,18 @@ export interface FileEntry {
 }
 
 /**
- * reads a file as text
- *
- * @param filePath path to the file
- * @param [options] configuration object
- * @param [options.dir] base directory
- * @return
+ * @name readTextFile
+ * @description Reads a file as text
+ * @param {string} filePath path to the file
+ * @param {FsOptions} [options] configuration object
+ * @param {BaseDirectory} [options.dir] base directory
+ * @return {Promise<string>}
  */
-async function readTextFile(filePath: string, options: FsOptions = {}): Promise<string> {
-  return await promisified({
+async function readTextFile(
+  filePath: string,
+  options: FsOptions = {}
+): Promise<string> {
+  return await promisified<string>({
     cmd: 'readTextFile',
     path: filePath,
     options
@@ -61,15 +64,18 @@ async function readTextFile(filePath: string, options: FsOptions = {}): Promise<
 }
 
 /**
- * reads a file as binary
- *
- * @param filePath path to the file
- * @param {Object} [options] configuration object
+ * @name readBinaryFile
+ * @description Reads a file as binary
+ * @param {string} filePath path to the file
+ * @param {FsOptions} [options] configuration object
  * @param {BaseDirectory} [options.dir] base directory
- * @return {Promise<int[]>}
+ * @return {Promise<number[]>}
  */
-async function readBinaryFile(filePath: string, options: FsOptions = {}): Promise<string> {
-  return await promisified({
+async function readBinaryFile(
+  filePath: string,
+  options: FsOptions = {}
+): Promise<number[]> {
+  return await promisified<number[]>({
     cmd: 'readBinaryFile',
     path: filePath,
     options
@@ -86,7 +92,10 @@ async function readBinaryFile(filePath: string, options: FsOptions = {}): Promis
  * @param [options.dir] base directory
  * @return
  */
-async function writeFile(file: FsTextFileOption, options: FsOptions = {}): Promise<void> {
+async function writeFile(
+  file: FsTextFileOption,
+  options: FsOptions = {}
+): Promise<void> {
   if (typeof options === 'object') {
     Object.freeze(options)
   }
@@ -145,7 +154,10 @@ function arrayBufferToBase64(buffer: ArrayBuffer): string {
  * @param [options.dir] base directory
  * @return
  */
-async function writeBinaryFile(file: FsBinaryFileOption, options: FsOptions = {}): Promise<void> {
+async function writeBinaryFile(
+  file: FsBinaryFileOption,
+  options: FsOptions = {}
+): Promise<void> {
   if (typeof options === 'object') {
     Object.freeze(options)
   }
@@ -170,7 +182,10 @@ async function writeBinaryFile(file: FsBinaryFileOption, options: FsOptions = {}
  * @param [options.dir] base directory
  * @return
  */
-async function readDir(dir: string, options: FsOptions = {}): Promise<FileEntry[]> {
+async function readDir(
+  dir: string,
+  options: FsOptions = {}
+): Promise<FileEntry[]> {
   return await promisified({
     cmd: 'readDir',
     path: dir,
@@ -224,7 +239,11 @@ async function removeDir(dir: string, options: FsOptions = {}): Promise<void> {
  * @param [options.dir] base directory
  * @return
  */
-async function copyFile(source: string, destination: string, options: FsOptions = {}): Promise<void> {
+async function copyFile(
+  source: string,
+  destination: string,
+  options: FsOptions = {}
+): Promise<void> {
   return await promisified({
     cmd: 'copyFile',
     source,
@@ -241,7 +260,10 @@ async function copyFile(source: string, destination: string, options: FsOptions 
  * @param [options.dir] base directory
  * @return
  */
-async function removeFile(file: string, options: FsOptions = {}): Promise<void> {
+async function removeFile(
+  file: string,
+  options: FsOptions = {}
+): Promise<void> {
   return await promisified({
     cmd: 'removeFile',
     path: file,
@@ -258,7 +280,11 @@ async function removeFile(file: string, options: FsOptions = {}): Promise<void> 
  * @param [options.dir] base directory
  * @return
  */
-async function renameFile(oldPath: string, newPath: string, options: FsOptions = {}): Promise<void> {
+async function renameFile(
+  oldPath: string,
+  newPath: string,
+  options: FsOptions = {}
+): Promise<void> {
   return await promisified({
     cmd: 'renameFile',
     oldPath,