+++ cli/tauri.js/api-src/window.ts
@@ -24,7 +24,4 @@ function open(url: string): void {
   })
 }
 
-export {
-  setTitle,
-  open
-}
+export { setTitle, open }