+++ cli/tauri.js/src/api/dependency-manager/types.ts
@@ -1,5 +1,6 @@
 export enum ManagementType {
   Install,
+  InstallDev,
   Update
 }