+++ cli/tauri.js/test/jest/fixtures/app-test-setup.js
@@ -6,15 +6,13 @@ const mockFixtureDir = path.resolve(__dirname, '../fixtures')
 module.exports.fixtureDir = mockFixtureDir
 
 function mockResolvePath(basePath, dir) {
-  return dir && path.isAbsolute(dir) ?
-    dir :
-    path.resolve(basePath, dir)
+  return dir && path.isAbsolute(dir) ? dir : path.resolve(basePath, dir)
 }
 
 module.exports.initJest = (mockFixture) => {
   jest.setTimeout(720000)
   jest.mock('helpers/non-webpack-require', () => {
-    return path => {
+    return (path) => {
       const value = require('fs').readFileSync(path).toString()
       if (path.endsWith('.json')) {
         return JSON.parse(value)
@@ -31,8 +29,8 @@ module.exports.initJest = (mockFixture) => {
       appDir,
       tauriDir,
       resolve: {
-        app: dir => mockResolvePath(appDir, dir),
-        tauri: dir => mockResolvePath(tauriDir, dir)
+        app: (dir) => mockResolvePath(appDir, dir),
+        tauri: (dir) => mockResolvePath(tauriDir, dir)
       }
     }
   })
@@ -63,7 +61,7 @@ module.exports.startServer = (onSuccess) => {
 
   function addResponse(response) {
     responses[response.cmd] = true
-    if (!Object.values(responses).some(c => c === null)) {
+    if (!Object.values(responses).some((c) => c === null)) {
       server.close(onSuccess)
     }
   }
@@ -83,7 +81,7 @@ module.exports.startServer = (onSuccess) => {
 
     if (req.method === 'POST') {
       let body = ''
-      req.on('data', chunk => {
+      req.on('data', (chunk) => {
         body += chunk.toString()
       })
       if (req.url === '/reply') {