+++ tauri/src/endpoints/notification.rs
@@ -2,7 +2,12 @@ use super::cmd::NotificationOptions;
 use serde_json::Value as JsonValue;
 use webview_official::Webview;
 
-pub fn send(webview: &mut Webview, options: NotificationOptions, callback: String, error: String) {
+pub fn send(
+  webview: &mut Webview<'_>,
+  options: NotificationOptions,
+  callback: String,
+  error: String,
+) {
   crate::execute_promise(
     webview,
     move || {
@@ -21,7 +26,7 @@ pub fn send(webview: &mut Webview, options: NotificationOptions, callback: Strin
   );
 }
 
-pub fn is_permission_granted(webview: &mut Webview, callback: String, error: String) {
+pub fn is_permission_granted(webview: &mut Webview<'_>, callback: String, error: String) {
   crate::execute_promise(
     webview,
     move || {
@@ -38,7 +43,7 @@ pub fn is_permission_granted(webview: &mut Webview, callback: String, error: Str
 }
 
 pub fn request_permission(
-  webview: &mut Webview,
+  webview: &mut Webview<'_>,
   callback: String,
   error: String,
 ) -> crate::Result<()> {